Variable blocks - ${Var} - now support some math and conditional operators. - Math operators are +, -, *, or / in form of ${Var + 3} etc - Conditional operator is C/etc ternary format: ${Var ? ifTrue : ifFalse} - Conditions can include =, !=, <, >, <=, >= like ${Var >= 0 ? pos : neg} - Else (:) is optional for above (returns Var itself if condition is false) - Only one operator per ${} (besides the ? and : after conditonal ones) - Nesting ${} blocks can combine multiple operators into complex statments

+++ Source/Profile.h
@@ -48,7 +48,7 @@ std::string getStr(const std::string& theSection,
 				   const std::string& theDefaultValue = "");
 int getInt(const std::string&, const std::string&, int = 0);
 bool getBool(const std::string&, const std::string&, bool = false);
-float getFloat(const std::string&, const std::string&, float = 0);
+double getFloat(const std::string&, const std::string&, double = 0);
 const PropertyMap* getSection(const std::string& theSectionName);
 const PropertyMap& getSectionProperties(const std::string& theSectionName);
 const SectionsMap& allSections();
